+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 
 - use `FetchDriver.svelte` so that boundaries are pre-fetched in the background
 - text in `XorSelector.svelte` can't be selectable anymore
+- for indicators with long `title` or `subtitle` the user can now expand the header to read them in full
 
 ## `@svizzle/ui` v0.6.0 (next)